Add files via upload
uploaded html with the JavaScript inside the html. also uploaded the css file. this project took alot of research as I have took many information and resources from the internet, specifically from youtube. as a beginner project, i felt this has been just the beginning of learning to be a self taught programmer. i plan to take what i've learned and do better and try harder on other projects.